Create Affine_Transformation.py
What does this program do?
Loads an image
Applies an Affine Transform to the image. This transform is obtained from the relation between three points. We use the function cv::warpAffine for that purpose.
Applies a Rotation to the image after being transformed. This rotation is with respect to the image center
Waits until the user exits the program